Fix listener on queue.pushed to process job immediately

+++ chancy/app.py
@@ -668,7 +668,7 @@ class Chancy:
         self,
         ref: Reference,
         *,
-        interval: int = 1,
+        interval: float = 1,
         timeout: float | int | None = None,
     ) -> QueuedJob | None:
         """
